backend/pyproject.toml
[project]
name = "langchain-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "edge-tts>=7.2.7",
    "elevenlabs>=2.36.1",
    "fastapi>=0.131.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "uvicorn>=0.41.0",
]
